requirement.txt
anyio==3.5.0
asgiref==3.4.1
backports.entry-points-selectable==1.1.1
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
distlib==0.3.3
Django==3.2.9
django-bootstrap4==21.2
django-crispy-forms==1.14.0
django-filter==21.1
django-resized==1.0.2
djangorestframework==3.13.1
fastapi==0.73.0
filelock==3.4.0
h11==0.13.0
humanfriendly==10.0
idna==3.3
importlib-metadata==4.10.1
Markdown==3.3.6
mixins==0.1.4
Pillow==9.0.1
platformdirs==2.4.0
pydantic==1.9.0
pyreadline3==3.4.1
python-decouple==3.6
pytz==2021.3
requests==2.27.1
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.1
sqlparse==0.4.2
starlette==0.17.1
typing_extensions==4.0.1
urllib3==1.26.8
uvicorn==0.17.4
virtualenv==20.10.0
zipp==3.7.0
